app/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-braces==1.14.0
django-comments-dab==2.6.1
django-crispy-forms==1.12.0
django-model-utils==4.1.1
django-social-share==2.2.1
django-star-ratings==0.9.1
django-widget-tweaks==1.4.8
Pillow==8.3.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
